+++ src/Data/Singletons.hs
@@ -39,8 +39,9 @@ module Data.Singletons (
   -- * Working with singletons
   KindOf, SameKind,
   SingInstance(..), SomeSing(..),
-  singInstance, pattern SingI,
-  withSingI, withSomeSing, singByProxy, demote,
+  singInstance, pattern Sing, withSingI,
+  withSomeSing, pattern FromSing,
+  singByProxy, demote,
 
   singByProxy#,
   withSing, singThat,
@@ -57,6 +58,9 @@ module Data.Singletons (
   singFun8,
   unSingFun1, unSingFun2, unSingFun3, unSingFun4, unSingFun5,
   unSingFun6, unSingFun7, unSingFun8,
+  -- $SLambdaPatternSynonyms
+  pattern SLambda2, pattern SLambda3, pattern SLambda4, pattern SLambda5,
+  pattern SLambda6, pattern SLambda7, pattern SLambda8,
 
   -- | These type synonyms are exported only to improve error messages; users
   -- should not have to mention them.
@@ -131,3 +135,42 @@ instance ShowSing k => Show (SomeSing k) where
 
 $(genDefunSymbols [''Demote, ''SameKind, ''KindOf, ''(~>), ''Apply, ''(@@)])
 -- SingFunction1 et al. are not defunctionalizable at the moment due to #198
+
+{- $SLambdaPatternSynonyms
+
+@SLambda{2...8}@ are explicitly bidirectional pattern synonyms for
+defunctionalized singletons (@'Sing' (f :: k '~>' k' '~>' k'')@).
+
+As __constructors__: Same as @singFun{2..8}@. For example, one can turn a
+binary function on singletons @sTake :: 'SingFunction2' TakeSym0@ into a
+defunctionalized singleton @'Sing' (TakeSym :: Nat '~>' [a] '~>' [a])@:
+
+@
+>>> import Data.Singletons.Prelude.List
+>>> :set -XTypeApplications
+>>>
+>>> :t 'SLambda2'
+'SLambda2' :: 'SingFunction2' f -> 'Sing' f
+>>> :t 'SLambda2' \@TakeSym0
+'SLambda2' :: 'SingFunction2' TakeSym0 -> 'Sing' TakeSym0
+>>> :t 'SLambda2' \@TakeSym0 sTake
+'SLambda2' :: 'Sing' TakeSym0
+@
+
+This is useful for functions on singletons that expect a defunctionalized
+singleton as an argument, such as @sZipWith :: 'SingFunction3' ZipWithSym0@:
+
+@
+sZipWith :: Sing (f :: a '~>' b '~>' c) -> Sing (xs :: [a]) -> Sing (ys :: [b]) -> Sing (ZipWith f xs ys :: [c])
+sZipWith ('SLambda2' \@TakeSym0 sTake) :: Sing (xs :: [Nat]) -> Sing (ys :: [[a]]) -> Sing (ZipWith TakeSym0 xs ys :: [[a]])
+@
+
+As __patterns__: Same as @unSingFun{2..8}@. Gets a binary term-level
+Haskell function on singletons
+@'Sing' (x :: k) -> 'Sing' (y :: k') -> 'Sing' (f \@\@ x \@\@ y)@
+from a defunctionalised @'Sing' f@. Alternatively, as a record field accessor:
+
+@
+applySing2 :: 'Sing' (f :: k '~>' k' '~>' k'') -> 'SingFunction2' f
+@
+-}